Fixes state transitions and adds res_active flag
Corrects a typo in the state machine enum and logic.

Adds a flag to track if a response from the autonomous
system is active, preventing multiple activations.
Adds a delay for testing and updates serial prints.

Also updates the mission select logic to respond to ASMS
or the res_active flag being set to HIGH.