Adds waiting ready state
Adds a new state for waiting before transitioning to the ready state.

Modifies the initial sequence to transition to the waiting ready state
instead of directly to the ready state.

Adds logic to read RPM from VCU can message, and ensures it's zero
before transitioning to finished state.

The solenoids now only activate when the state is emergency to prevent them to activate when the acu is turned on.