Adds manual mode functionality
Implements a manual mode for the autonomous system, allowing direct control.

This change introduces a new `STATE_MANUAL` state, which can be triggered via a CAN message. When active, it deactivates solenoids, resets flags, and sets the autonomous system to an off state.

Also resets the inital sequence variables when ASMS flag is low, fixing a bug.